A ring arrives and leaves on the Fade, and a played node always wears one
The Gate (#395) is a threshold over a live measurement, so a partial breathing
across it took its node's whole annulus with it every time it crossed, and a
node still sounding could lose its ring part way through the release the rest
of it was still drawing.

Two rules, which are one rule read from two ends — the audio ring is a LAYER OF
A NODE, and every other layer of a node arrives and leaves on the Note
section's Fade:

- **A ring comes and goes on the Fade.** `RingFade` carries the gate's answer
  per bucket of the analyzer's grid across frames, on the note envelope
  (`ViewConfig::envelope`, so the Fade param and the Fade curve both reach it).
  A level breathing across the bar is now a ring dimming and coming back rather
  than one flickering.
- **A node the keys have lit wears its ring whatever the gate says**, at the
  level the rest of the node is drawn at — so the ring leaves exactly with the
  note instead of a layer dropping out mid-release. The Gate is thereby a
  question about the nodes NOBODY is playing, which is where a ring at every
  node was worth holding back.

`NodeInstance::audio_ring` is a level rather than a bit, and reaches the shader
as one (`@location(11)`), where it multiplies the ring's COVERAGE — so what
shows through a ring on its way in is the octave layer under it, not a reading
of a quieter spectrum. It still lands exactly on 0, so an idle node whose ring
has gone stops being shipped at all.

`Envelope::carried` is the new primitive: a LEVEL carried `dt` further along the
one curve `attack`/`release` already run. A level and not a moment, because the
gate has no note-on to restart from — it can cross back before the fade it
started has landed, and a transition restarted from a stamp would jump. Stepping
is exact, so a render at 60 fps and a pane at 144 walk one transition, and a
`dt` of 0 holds — which is what keeps the two lattices an editor frame draws
(docked pane, Video preview) from stepping the fade twice.

The state lives in `SharedState`, which is the whole of what the offline
renderer carries between frames; `rendering_the_same_take_twice_is_byte_identical`
still holds.

## Tests

- `harmonigraph-core`: a carried level walks the curve the attack does at every
  shape, steps compose exactly, it lands exactly on its target, and every
  poisoned input steps rather than hangs.
- `harmonigraph-scene`: a ring comes and goes on the Fade and turns around from
  where it got to, a Fade of 0 is the gate itself, one frame drawn twice steps
  it once, a played node rings at the gate's ceiling, and its ring leaves on the
  note's own fade.
- `harmonigraph-ui`: a tone arriving rings its class half way through the Fade
  and fully after it, end to end from real audio.
- `harmonigraph-render` (GPU): a ring part way through its fade is every pixel
  between the two ends, and nothing outside the annulus moves.

`./ci.sh` green.

## Compat

No new keys and no reinterpretation: the Fade the ring runs on is the Fade the
notes already run on.

+++ crates/harmonigraph-render/src/lib.rs
@@ -404,19 +404,18 @@ struct GpuInstance {
     /// sheet), y = knockout gutter width in uv units (0 on the home sheet).
     /// See `NodeInstance::scale` / `::gutter`.
     sevens: [f32; 2],
-    /// Whether this node draws the audio ring: 1 where at least one of its
-    /// wedges reaches the gate, 0 where none does
+    /// How much of the audio ring this node wears, 0..=1: the gate's answer for
+    /// its wedges carried on the note Fade, floored by the node's own envelope
     /// (`NodeInstance::audio_ring`).
     ///
-    /// A DECISION and not the node's own peak level with the gate beside it in
-    /// the uniforms, though there is a free slot there for one: the rule is
-    /// "the loudest wedge reaches the gate", the levels and the wheel it is
-    /// measured over both live on the CPU, and splitting the comparison across
-    /// the bus would leave two places able to disagree about which nodes ring.
-    /// A float and not a `u32` because a vertex attribute has no bool and the
-    /// shader tests it as a level; it is 0 or 1 all the same, the gate being a
-    /// question about whether a ring is on the screen (see
-    /// `SpectralPaint::gate`).
+    /// A DECISION already taken and not the node's own peak level with the gate
+    /// beside it in the uniforms, though there is a free slot there for one:
+    /// the rule is "the loudest wedge reaches the gate", the levels and the
+    /// wheel it is measured over both live on the CPU, and splitting the
+    /// comparison across the bus would leave two places able to disagree about
+    /// which nodes ring. What crosses is where that decision has GOT to, which
+    /// is a level because a ring arrives and leaves on the Fade like every
+    /// other layer of a node (see `harmonigraph_scene::RingFade`).
     ring: f32,
 }
 
@@ -749,7 +748,7 @@ impl LatticeCallback {
                 melody_color: n.melody_color.to_array(),
                 bass_color: n.bass_color.to_array(),
                 sevens: [n.scale, gutter],
-                ring: f32::from(u8::from(n.audio_ring)),
+                ring: n.audio_ring,
         };
 
         let split = order
@@ -774,14 +773,16 @@ impl LatticeCallback {
         //
         // The audio RING is why this is not a property of the node alone: the
         // ring is a window onto the spectrum rather than a level a node
-        // carries, so it takes BOTH the layer being on and this node having
-        // been let through the gate (`Scene::gate_audio_rings`) for the node to
-        // owe an annulus. With the gate at its floor that is every node in the
-        // window, silence included — the ungated picture, and what "the ring
-        // reads raw" costs; dialled up, an idle node with nothing sounding at
-        // it goes back to shipping nothing at all. The shader's idle branch
-        // pays the rest per fragment, keeping an otherwise idle node to the
-        // ring's own annulus.
+        // carries, so it takes BOTH the layer being on and this node wearing
+        // some of the ring (`Scene::wear_audio_rings`) for the node to owe an
+        // annulus. With the gate at its floor that is every node in the window,
+        // silence included — the ungated picture, and what "the ring reads raw"
+        // costs; dialled up, an idle node with nothing sounding at it goes back
+        // to shipping nothing at all, ONCE ITS FADE HAS RUN OUT — the level
+        // reaches exactly 0 rather than approaching it, so a ring on its way
+        // out is shipped for exactly as long as it is drawn. The shader's idle
+        // branch pays the rest per fragment, keeping an otherwise idle node to
+        // the ring's own annulus.
         let ringing = scene.spectral.ring_draws();
         let paints = |g: &GpuInstance| {
             (ringing && g.ring > 0.0)
